async cleanup

+++ dinner_project/frontend/src/_auth/axiosConfig.js
@@ -2,6 +2,8 @@ import axios from 'axios';
 
 const API_URL = "http://127.0.0.1:8000/api/";
 
+let refreshInProgress = false;
+
 const axiosInstance = axios.create({
     baseURL: API_URL,
     timeout: 5000,
@@ -16,7 +18,7 @@ axiosInstance.interceptors.response.use(
     function (response) {
         //This triggers for anything where status code is within 2xx
         return response;
-    }, function (error) {
+    }, async function (error) {
 
         const originalRequest = error.config;
 
@@ -31,24 +33,33 @@ axiosInstance.interceptors.response.use(
             return Promise.reject(error);
         }
 
-        return axios.post(API_URL + 'token/refresh/', {
-            refresh: localStorage.getItem('refresh_token')
-        })
-        .then(function (response) {
-            localStorage.setItem('access_token', response.data.access);
-            localStorage.setItem('refresh_token', response.data.refresh);
-        
-            axiosInstance.defaults.headers['Authorization'] = 'JWT ' + response.data.access;
-            originalRequest.headers['Authorization'] = 'JWT ' + response.data.access;
-        
+        if (!refreshInProgress) {
+            try {
+                refreshInProgress = true;
+                const response = await axios.post(API_URL + 'token/refresh/', {
+                    refresh: localStorage.getItem('refresh_token')
+                });
+                refreshInProgress = false;
+                localStorage.setItem('access_token', response.data.access);
+                localStorage.setItem('refresh_token', response.data.refresh);
+
+                axiosInstance.defaults.headers['Authorization'] = 'JWT ' + response.data.access;
+                originalRequest.headers['Authorization'] = 'JWT ' + response.data.access;
+                return axiosInstance(originalRequest);
+            }
+            catch (error) {
+                refreshInProgress = false;
+                localStorage.clear();
+                window.location.href = '/login';
+                console.log(error);
+                return Promise.reject(error);
+            } 
+        }
+        else {
+            //TODO: add 'sleep'
             return axiosInstance(originalRequest);
-        })
-        .catch(function (error) {
-            localStorage.clear();
-            window.location.href = '/login';
-            console.log(error);
-            return Promise.reject(error);
-        }); 
+        }
+
     }
 );
